change connectionPending to long because that's the type returned by the underlying data structure

+++ Palvelukartta.m
@@ -165,7 +165,7 @@ NSString* ctostr(NSURLConnection* c) {
     [data appendData:newdata];
 }
 
-- (int) connectionsPending {
+- (long) connectionsPending {
     return [remainingConnections count];
 }
